refactor: DTOSS-9159 - split file-level validation into four separate classes
